+++ lib/entities/deities.dart
@@ -1,0 +1,1 @@
+export '../src/entities/deities/deity.dart' show Deity;