+++ lib/src/generators/deities/attributes/deity_positive_attribute_generator.dart
@@ -1,0 +1,7 @@
+import '../../base/list_item_generator.dart';
+import 'deity_attributes_data.dart';
+
+/// A class that generates positive attributes for deities
+class DeityPositiveAttributeGenerator extends ListItemGenerator<String> {
+  DeityPositiveAttributeGenerator() : super(deityPositiveAttributes);
+}