+++ lib/subtypes/deities.dart
@@ -1,0 +1,7 @@
+export '../src/subtypes/deities/deity_type.dart' show DeityType;
+export '../src/subtypes/deities/deity_manager.dart' show DeityManager;
+
+export '../src/subtypes/deities/gods/god.dart' show God;
+export '../src/subtypes/deities/demigods/demigod.dart' show Demigod;
+export '../src/subtypes/deities/demon_lords/demon_lord.dart' show DemonLord;
+export '../src/subtypes/deities/archangels/archangel.dart' show Archangel;